third_party/blink/web_tests/platform/generic/storage/websql/transaction-callback-isolated-world-expected.txt

ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the transaction callback is called in the right world.
